雅琳-修註冊/登入

+++ src/main/java/foodelicious/member/model/TotalDao.java
@@ -65,6 +65,6 @@ public class TotalDao {
 
 		Account account = query.uniqueResult();
 		session.close();
-		return account.getAccount_id();
+		return (long) account.getAccount_id();
 	}
 }
